Change the core->ltx function to produce a string
The final transformation of a template should produce a string. If you
produce that as the final output, you should be able to place that right
into a file with no problems.

The core tags should transform directly into a string. I wrote up a
little proof to show that the core->ltx function ought to do that.