Made progress on fixing list processing
Instead of splitting at bullets preceeded by a sequence of whitespace
and newlines, I'm splitting at a newline that's followed by whitespace
and a bullet.

By doing things this way, the preceeding whitespace of a line can stay
with the bullet that comes after it. I can stitch together the
whitespace and bullet afterward. With this change, the lists now fit
into the model that I was working with, where we have a sequence of
whitespace, bullet, newline, whitespace, bullet, ... where bullets can
be strings that start with some regexp pattern, or tags of a certain
name (like 'i).