Change macros to allow lists of core? for each arg
Before, macros took a list of strings and core tags as their arguments.
The problem with this is that often enough, a list of strings (or
txexpr-element?) is going to be the argument to the macro. In order to
allow a list of strings, I allowed each argument to be a list of strings
or core tags.